Move samples into pantry.
