+++ src/geovista/examples/point_cloud/from_points__orca_cloud_eqc.py
@@ -33,7 +33,7 @@ from __future__ import annotations
 
 import geovista as gv
 from geovista.pantry.data import um_orca2_gradient
-from geovista.samples import ZLEVEL_SCALE_CLOUD
+from geovista.pantry.samples import ZLEVEL_SCALE_CLOUD
 import geovista.theme
