+++ src/geovista/geoplotter.py
@@ -22,6 +22,13 @@ from pyproj import CRS
 import pyvista as pv
 from pyvista.core.utilities import helpers
 
+from geovista.pantry.samples import (
+    LFRIC_RESOLUTION,
+    REGULAR_RESOLUTION,
+    lfric,
+    regular_grid,
+)
+
 from .bridge import Transform
 from .common import (
     GV_FIELD_ZSCALE,
@@ -55,7 +62,6 @@ from .gridlines import (
     create_parallels,
 )
 from .raster import wrap_texture
-from .samples import LFRIC_RESOLUTION, REGULAR_RESOLUTION, lfric, regular_grid
 from .transform import transform_mesh
 
 if TYPE_CHECKING:
@@ -95,7 +101,7 @@ def _lfric_mesh(
     ----------
     resolution : str, optional
         The resolution of the LFRic unstructured cubed-sphere. Defaults to
-        :data:`geovista.samples.LFRIC_RESOLUTION`.
+        :data:`geovista.pantry.samples.LFRIC_RESOLUTION`.
     radius : float, optional
         The radius of the sphere. Defaults to :data:`geovista.common.RADIUS`.
 
@@ -287,7 +293,7 @@ class GeoPlotterBase:
         resolution : str, optional
             The resolution of the cubed-sphere to generate as the base layer,
             which may be either ``c48``, ``c96`` or ``c192``. Defaults to
-            :data:`geovista.samples.LFRIC_RESOLUTION`. Alternatively, generate a
+            :data:`geovista.pantry.samples.LFRIC_RESOLUTION`. Alternatively, generate a
             regular grid using a format of ``rN``, where ``N`` is the number of cells
             in latitude, and ``N * 1.5`` cells in longitude. When adding a base layer
             to a projection, the default is to use a regular grid with resolution
