+++ tests/samples/__init__.py
@@ -3,4 +3,4 @@
 # This file is part of GeoVista and is distributed under the 3-Clause BSD license.
 # See the LICENSE file in the package root directory for licensing details.
 
-"""Unit-tests for :mod:`geovista.samples`."""
+"""Unit-tests for :mod:`geovista.pantry.samples`."""
